feat(orders): quantify orders in grams, not Toman price
Wholesale gold is measured by weight. Order total = Σ(product weight_grams × qty);
each line snapshots unit_weight_grams. Legacy unit_price/price columns kept for old
records but no longer populated. Stats total_value now = total grams. Migration 0013.

+++ backend/app/api/v1/admin_stats.py
@@ -79,7 +79,7 @@ async def stats(db: AsyncSession = Depends(get_db)):
         "orders_week": await count_since(week),
         "orders_month": await count_since(month),
         "total_orders": int(total_orders),
-        "total_value": int(total_value),
+        "total_value": float(total_value),  # total grams sold
         "unread": int(unread),
         "conversion_rate": round(confirmed / total_orders, 3) if total_orders else 0,
         "orders_by_day": orders_by_day,
